requirements.txt
# MediRisk Medical Prediction System Requirements
# ================================================

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data Science and Machine Learning
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization (Apple Silicon uyumlu versiyonlar)
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.0.0

# Data Processing
scipy==1.11.3
openpyxl==3.1.2

# Web Development
Jinja2==3.1.2
Werkzeug==2.3.7
MarkupSafe==2.1.3

# API and Validation
marshmallow==3.20.1
jsonschema==4.19.1

# Security
cryptography==41.0.7

# Logging and Monitoring
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.0
pytest-cov==4.1.0
httpx>=0.25.0  # For testing FastAPI

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Jupyter Notebook Support
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.26.0
ipywidgets==8.1.1

# Optional: Database Support (for future enhancements)
SQLAlchemy==2.0.23
alembic>=1.12.0  # Database migrations

# Optional: Async Support (for future enhancements)
celery==5.3.4
redis==5.0.1

# Development Tools
black==23.9.1
flake8==6.1.0
mypy==1.7.1

# Production Server (FastAPI compatible)
uvicorn[standard]>=0.24.0

# Memory profiling (optional)
memory-profiler==0.61.0

# Progress bars for data processing
tqdm==4.66.1

# LLM Integration for Report Enhancement
aiohttp>=3.8.0
requests>=2.28.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=0.19.0
orjson>=3.8.0
